services/pascal-legacy/requirements.txt
psycopg2-binary==2.9.7
openpyxl==3.1.2
